Assignment Complete (With updated features)
Added the onclick for all users to see their timeline. Didn't have enought time on the day of the deadline because I was to focused on learning about the DOM's fundalmental since working with the DOM and manipulate the DOM with JS is something totally new to me.

Now a day after the deadline I was working on this agin all day and I finally start to really understand the DOM & to manipulate it with Javascript & jQuery with this update.

I know it's after the deadline but please have a look at it. I would really aprriciate if you guys can review it.

There's not a whole lot of CSS since I really focues on manipulating elements. So once I finish my last deadline for module 3, I'll come back to this to get more practice. But I'm quite happy that I was able to add a on-click to re-direct to a user's timeline.

Thank you and have a nice day